requirements.txt
fastapi>=0.110,<0.112
uvicorn[standard]>=0.29,<0.31
feedparser>=6.0,<6.1
beautifulsoup4>=4.12,<4.13
requests>=2.31,<2.33
google-generativeai>=0.5,<0.7
python-dotenv>=1.0,<1.1
